Fix DrawStyleParentBackgroundEx wrong portion of background used
DrawStyleParentBackgroundEx used the coordinates of the parentwindow (LPoint) to copy from LBuffer device context, thats wrong to copy from the bitmap dc the source position must be 0, 0.  For the WM_ERASEBKGND you must change the Origin of the bitmap devicecontext so that the right part of the background is painted into to bitmap.

(DrawStyleParentBackground may be removed later, because its complete wrong and doesn't paint anything to DC !)