Fix drawing Buttons in Designmode, and also for Delphi ObjectInspektor align popup
Detour_UxTheme_GetThemeColor: Reorganized code instead that every TMT_TEXTCOLOR is answered with clWindowText. First give a try to each case if this doesn't set the pColor to something different as clNone - stay at clWindowText. For VSCLASS_MENU we get ask for the TMT_TEXTCOLOR and for the TMT_FILLCOLOR of the same state.

Replaced calls to DrawStyleParentBackground with DrawStyleParentBackgroundEx after fixing it. -> this also fixes drawing issues with TButton in Designmode.

UxTheme_Button: Reverted the drawing on a temporary bitmap with a simple transparent color - because the fill and transparentcolor clBtnFace is also the color used of some style to draw the button surface .. which becomes in that case transparent and not visible.
If the button itself has rounded corners - these are drawn with semitransparency in the StyleService - so if a temporary bitmap is used it must be set to pf32bit with AlphaFormat instead of a transparent color.
After replacing not working DrawStyleParentBackGround with my modifed DrawStyleParentBackgroundEx it seems to work.